# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nokta Academy API",
    "description": "Educational management system API for schools and academies.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8081", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List users",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "User list" } }
      }
    },
    "/api/attendance": {
      "get": {
        "tags": ["Academic"],
        "summary": "Attendance records",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Attendance data" } }
      }
    },
    "/api/finance": {
      "get": {
        "tags": ["Finance"],
        "summary": "Finance overview",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Finance data" } }
      }
    },
    "/api/books": {
      "get": {
        "tags": ["Academic"],
        "summary": "Book inventory",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Books" } }
      }
    },
    "/api/audit": {
      "get": {
        "tags": ["System"],
        "summary": "Audit log",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Audit entries" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}